refactor(rename files): rename files to use component name
Rename files from using Index to using the name of the component. i.e if a compoennt name is
CheckBox the subfo
