+++ src/components/checkBox/checkBox.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import CheckBox from ".";
+import CheckBox from "./CheckBox";
 
 export default {
   title: "CheckBox",
